FINAL VERSION/requirements.txt
ctfcli
PyYAML